♻️ [Refactoring]: Lexer<'a> を lexer.rs に直書きに統合し cursor.rs を削除
公開 path を `crate::lexer::cursor::Lexer` から `crate::lexer::Lexer` に短縮。
parent module（lexer.rs）に lexer 本体を、サブモジュール（byte_kind / eol /
token）に「素材」を置く Rust 慣用パターン（std::vec::Vec と vec.rs 等）に
合わせる。型名と中身（cursor.rs に Lexer<'a>）の名前不一致も解消。

Refs #400
